functions/repository-indexer/requirements.txt
PyGithub==1.55
firebase-admin==6.1.0
python-dotenv==0.19.2
pytest==7.4.0
